wv: WVCTCS Banner SSB cluster scraper (5 new colleges) + wvncc ceiling
Investigation found that 5 of WV's 6 uncovered colleges share the WVNET Banner
SSB 9 backend (xe<code>prod.wvnet.edu) already used for Mountwest — their public-
facing *.edu portals (Ellucian Experience SSO / sgcaptcha WAF / JS schedule) gate
registration, not the backend class search. Prior config marked all six manual-only;
that was the front-door verdict.

- Rename scripts/wv/scrape-mountwest.ts -> scrape-wvctcs-banner.ts; HOSTS now maps
  6 colleges (mountwest + blueridge/bridgevalley/newriver/pierpont/southern) to their
  verified-public WVNET Banner hosts. Same scrapeBannerSsbState flow.
- Wire the renamed scraper in lib/states/wv/config.ts; drop the 5 manual-only markers.
- Document wvncc as a course ceiling (its WVNET host has no SSB/8 module; CollegeScheduler
  tenant stale to Fall 2023; Pathify SAML) — genuinely no public sections.

Data for the 5 new colleges lands in a follow-up commit on this branch.
